Remove VERSION variable
We now store it's value in buildPackage.sh, which is the only place it was used. It was always a bit daft that we required it to be specified manually, but even worse, it was breaking the PySide build because setup.py was treating it as an override for a command line argument.