Alembic : Update to version 1.7.5
This matches the requirements of VFX Platform 2018. It also fixes a linking error whereby libAlembic doesn't link to libhdf5, and instead expects the libAlembic client library to ensure that it is linked. That approach has been working for us until now, but isn't working on Travis for the Cortex tests, where randomly libAlembic is getting loaded first and then complaining about missing HDF5 symbols. See the following relevant commits and issues :

https://github.com/alembic/alembic/commit/69f15e899a70ebe9cbb72d14fbf9d99ca174f4e8
https://github.com/alembic/alembic/commit/61370ccf4cf634eb56fc8ee8d71f17902f41f47a
https://github.com/alembic/alembic/issues/126