Boost : Build without ICU support
Previously we were omitting ICU support from dependencies builds purely by virtue of the fact I didn't have the headers on my machine. An upgrade to CentOS7 appears to have left me with the headers, so we must explicitly disable ICU support to get back to the status quo. Otherwise we end up with a dependency on a particular version of `libicu.so`, and we can't guarantee that it will be available wherever Gaffer will be run (Travis has a different version, for instance).

Really I need to learn about these container things that @donboie keeps telling me about, so this is less likely to happen.