FreeType : Explicitly build without harfbuzz
This matches previous builds where we have just been fortunate to build without harfbuzz. From the sounds of it, harfbuzz might be a good thing, but we can't rely on it being on any particular machine - if we did want to include it, we'd have to include it in the dependencies.