Add generic config driven build script
Many of our build scripts share the same boilerplate, and we can avoid repeating that by moving all the logic regarding archives etc into a single config driven script.

This new script also downloads archives from the internet rather than expecting them to be hosted in the repo. The repo is becoming utterly unwieldy with the historical weight of all the archives, so we need to adopt a smarter approach.