Revert "TBB : Enable community preview features"
This reverts commit f01b0d315a19f7300c8d676374cb1d8e1cbdde14.

I hadn't realised that doing this meant that libtbb_preview.so was built _instead_ of libtbb.so. I had assumed it was an additional library, not a replacement. This is breaking the subsequent build for OpenVDB, and is probably a bad idea in general if we want to keep our code compatible with whatever TBB ships with various DCCs with their own unique interpretation of the VFX Platform.

Fixes #122.