Add LibFFI 3.3, and use `--with-system-ffi` for Python
Python 2 used to bundle the source for libffi and use that by default. Python 3 always uses the system libffi on Linux, no matter what you ask it to do, so we need to control the library version to guarantee compatibility on different distros. We are also using own own ffi for Python 2 and on Mac as well for consistency.

Fixes #163.