OpenSSL : Include libraries in manifest
Before we upgraded to 1.1.1.h, the default was to make static libs, so we didn't need to ship them. When we upgraded, OpenSSL switched to building shared libs, but I didn't add them to the manifest. So we ended up building Python modules with dependencies on libs we weren't packaging.