OIIO : Update to 2.2.11.1
This also required the following updates :

- LibTIFF to 4.1.0, because OIIO no longer supports older versions.
- OpenShadingLanguage to 1.11.11.0, because the old version doesn't support the new OIIO.
- LLVM to version 10.0.1, because OpenShadingLanguage wants a more recent version.

Appleseed
---------

We need to patch the shader include path because Appleseed unwisely modifies `stdosl.h` rather than simply add extension headers of its own. This makes shader compilation fragile, particularly because since we last updated OSL, it has mucked about with the search process for `stdosl.h` _twice_. The latest version is this :

https://github.com/AcademySoftwareFoundation/OpenShadingLanguage/pull/956

So we now use `-I` to redirect `oslc` to where we want it to find `stdosl.h`, rather than relying on executable-relative search as before.