LLVM : Set GCC_INSTALL_PREFIX correctly
Otherwise we get problems finding C++ headers when OSL tries to compile with Clang. With thanks to boberfly for doing the hard work of figuring this out.

Also updated `build.py` to run via Python 3, so we can use `shutil.which()` instead of rolling our own. Python 2 is dead, and while we still need to support it in Gaffer itself for a bit longer, we can at least update our build toolchain.

Fixes #193