Cycles : Add version 3.1.1
The patch for `object.h` is a very strange one. Without the patch we get this crash :

```
/home/john/dev/gafferDependencies/Cycles/working/cycles-3.1.1/src/graph/node.cpp:72: void ccl::Node::set(const ccl::SocketType&, bool): Assertion `input.type == SocketType::BOOLEAN' failed.
0  0x00007ffff6f11387 in raise () from /lib64/libc.so.6
1  0x00007ffff6f12a78 in abort () from /lib64/libc.so.6
2  0x00007ffff6f0a1a6 in __assert_fail_base () from /lib64/libc.so.6
3  0x00007ffff6f0a252 in __assert_fail () from /lib64/libc.so.6
4  0x00007fff2724379c in ccl::Node::set(ccl::SocketType const&, bool) () from /home/john/dev/build/gaffer62Cycles/lib/libGafferCycles.so
5  0x00007fff26c37f78 in ccl::Object::set_geometry (this=0x7fff06bfd8d0, value=0x7fffd0812c00) at /home/john/dev/build/gaffer62Cycles/cycles/include/scene/object.h:54
6  0x00007fff26d59cb8 in ccl::Geometry::compute_bvh(ccl::Device*, ccl::DeviceScene*, ccl::SceneParams*, ccl::Progress*, int, int) () from /home/john/dev/build/gaffer62Cycles/lib/libGafferCycles.so
```

This is because unless a definition of `Geometry` is available at the time it is defined, `set_geometry()` will end up implicitly casting `Geometry *` to bool and calling `Node::set( ccl::SocketType const&, bool )`. It's the same reason that `scene/particles.h` is included explicitly, as described in the comment for that. The mystery is how this has ever worked for _anyone_. Perhaps when building for Blender, `geometry.h` is included via some other path?

cycles

changes