OpenColorIO : Build required dependencies explicitly
The latest version of OpenImageIO has changed how it searches for OCIO, and now gets very upset if the lib has been moved from the location it was originally built in. So we can't use our trick of moving from `lib64` to `lib`. But as documented in the previous comments, we couldn't use `CMAKE_INSTALL_LIBDIR` to get it right in the first place, because that completely broke OCIO's automatic builds for Expat, YAML-CPP and PyString.

So instead, we now build all those dependencies ourselves, tell OCIO never to build anything else, and use `CMAKE_INSTALL_LIBDIR` like we wanted to all along. This does have one real benefit : we now correctly credit these third-third-party dependencies in the Gaffer documentation.