Fmt : Add version 9.1.0
We were getting this via OpenImageIO before, with the headers installed in `include/OpenImageIO/detail/fmt`. With the upgrade to OpenImageIO 2.4, this install no longer included the additional header needed for `dynamic_format_arg_store`, so we now build and install it ourselves. This will also allow us to take advantage of `fmt` more widely in Gaffer, in preparation for an eventual C++20 world where we can replace it with `std::fmt`.