Cycles : Update to version 3.4
This is necessary to get a [fix](https://github.com/blender/cycles/commit/d689b82704f542b6b8427d3a26204dfc6a3927f0) for compatibility with OSL 1.12, which changed behaviour in a way which stopped various Cycles shaders getting `N`. This in turn broke Gaffer unit tests, because Gaffer's default shader is a facing ratio that relies on `N`.

The Hydra render delegate isn't compatible with USD 23.02, so it has been disabled. It's not needed in Gaffer. The delegate now also depends on `libexpoxy` which would add yet another dependency anyway, and also a requirement for the `meson` build system in our build containers. I don't think the delegate is relevant enough to Gaffer to warrant this complexity at the moment.