Cycles : Enable CUDA and Optix builds
The latter requires the `OPTIX_ROOT_DIR` environment variable to have been set in the build environment, but that is not our responsibility here. We want `GafferHQ/dependencies` to be as build-environment-agnostic as it can be, so other folks can use it to make their own builds in their own environments.