Python : Fix linking against openssl11
The approach taken in 9e3fdbc71e7b3b6a08ad993c82b11556fb2526df resulted in the `python` binary also linking against libssl.so.1.1, which then prevented Python from running on systems without it.

So we back out of that and instead adjust the build container in https://github.com/GafferHQ/build/pull/39 to allow `configure` to automatically find the openssl11 headers and libraries. This then leads to the `_ssl` and `_hashlib` modules linking against libssl.so.1.1, but the `python` binary untouched.