Build : Allow arch specific overrides
Allows build setting to be modified in an arch specific way.
This can be useful on aarch64 where the required versions of
dependencies may not build correctly without additional
settings, or even be supported at all.

(Discovered building on aarch64 on Alma 9, where OpenPGL only
supports aarch64 in 0.6.0, which includes breaking changes
that are not compatible with Cycles.