Jemalloc : Enable `tls_model("initial-exec")`
This avoids the potential of infinite recursion with dynamic TLS on modern glibc versions, as noted here: https://lists.gnu.org/archive/html/info-gnu/2024-01/msg00017.html

```
  The dynamic linker calls the malloc and free functions in more cases
  during TLS access if a shared object with dynamic TLS is loaded and
  unloaded.  This can result in an infinite recursion if a malloc
  replacement library or its dependencies use dynamic TLS instead of
  initial-exec TLS.
```

The patch is based on the below commit for Jemalloc 4.0.0, though `configure` is being patched instead of `configure.ac` as we call `configure` directly.

https://github.com/jemalloc/jemalloc/commit/f1cf3ea4753260d37c9a43463bae2140e00e16f0

This suggests that `initial-exec` was the intended default but the test has been long broken.