OpenColorIO : Update to version 2.3.2
We're put into a bit of a odd corner with Minizip's defined value of `MZ_VERSION_BUILD`, which OCIO relies on for compatibility with both Minizip 3.x and 4.x.

In Minizip 3.0.8 and above this is defined as an octal literal with apparently? hexadecimal values. Recent releases of Minizip 4 correctly prepend the hexadecimal value with 0x, but OCIO's version checks haven't caught up. So we're left either patching `MZ_VERSION_BUILD` to a valid octal value or updating to Minizip 4.x and patching OCIO's `#if MZ_VERSION_BUILD >= 040000` checks to `#if MZ_VERSION_BUILD >= 0x40000`. Of the two choices, I've gone with the one requiring the smallest patch.