Qt/PySide : Update to version 6.5.5
qtpurchasing is no longer included with Qt6 so we no longer need to skip it.
We skip additional Qt modules that will not build due to us skipping qtdeclarative (it seems we're required to manually specify them all).
moc, rcc & uic are now installed to `{buildDir}/libexec` so we copy them over to `{buildDir}/bin`. There is an option to change the install location to `{buildDir}/bin`, but PySide appears to expect them in `libexec` and doesn't build pyside6-uic as a result.
We patch PySide to remove the dependency on `packaging`. It is only used to test whether we're on >= Qt 5.7 and would otherwise require updating the build container or pip installing `packaging` to Gaffer's Python.